Import version 0.5.0
Changelog
- rewrite of the core (no new features)

The "core" of the whole script is now in lib/ffrk-proxy.js, bin/app.js is now just a starter script, which can be modified. The core is now using events, which can be used to manipulate data at specific stages. For now there is one event "battleInitData". The launcher script is using the event to change the buddy and enemy data.