Update enemy.js
Initial attempt to fix and leave Vulnerability on enemies - easier to confirm if target conditions are met. This is workaround code for the current setup. May be tidied up in the future, but currently I have nothing in the way of getting away from the switch condition to display elemental weakness. Does not display all due to size of some entries.